+++ apps/dex/src/compounds/ImportModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ImportModal";
